e-reading-server/requirements.txt
astroid==2.1.0
cachetools==3.1.0
dateparser==0.7.1
dj-database-url==0.5.0
Django==2.1.7
django-heroku==0.3.1
djangorestframework==3.9.1
gitdb2==2.0.5
GitPython==2.1.11
google-api-python-client==1.7.8
google-auth==1.6.3
google-auth-httplib2==0.0.3
gunicorn==19.9.0
httplib2==0.12.1
humanize==0.5.1
isort==4.3.4
lazy-object-proxy==1.3.1
lint==1.2.1
maya==0.6.1
mccabe==0.6.1
nltk==3.4
pendulum==2.0.4
psycopg2==2.7.7
psycopg2-binary==2.7.7
pyasn1==0.4.5
pyasn1-modules==0.2.4
pylint==2.2.2
python-dateutil==2.8.0
pytz==2018.9
pytzdata==2019.1
regex==2019.4.14
rsa==4.0
singledispatch==3.4.0.3
six==1.12.0
smmap2==2.0.5
snaptime==0.2.4
sqlparse==0.3.0
typed-ast==1.3.1
tzlocal==1.5.1
uritemplate==3.0.0
whitenoise==4.1.2
wrapt==1.11.1
yapf==0.26.0
